Move blog tests into separate files
