+++ blog/tests/test_models.py
@@ -1,21 +1,10 @@
 from datetime import timedelta
 
-from django.core.urlresolvers import reverse
 from django.test import TestCase
 from django.test.utils import captured_stderr
-from django.utils import timezone
-from django.utils.timezone import now
 
-from blog.forms import EntryForm
-
-from .models import Entry, Event
-
-
-class DateTimeMixin(object):
-    def setUp(self):
-        self.now = timezone.now()
-        self.yesterday = self.now - timedelta(days=1)
-        self.tomorrow = self.now + timedelta(days=1)
+from . import DateTimeMixin
+from ..models import Entry, Event
 
 
 class EntryTestCase(DateTimeMixin, TestCase):
@@ -86,59 +75,3 @@ class EventTestCase(DateTimeMixin, TestCase):
 
         self.assertQuerysetEqual(Event.objects.future(), ['c', 'd'], transform=lambda event: event.headline)
         self.assertQuerysetEqual(Event.objects.past(), ['b', 'a'], transform=lambda event: event.headline)
-
-
-class ViewsTestCase(DateTimeMixin, TestCase):
-    def test_no_past_upcoming_events(self):
-        """
-        Make sure there are no past event in the "upcoming events" sidebar (#399)
-        """
-        # We need a published entry on the index page so that it doesn't return a 404
-        Entry.objects.create(pub_date=self.yesterday, is_active=True)
-        Event.objects.create(date=self.yesterday, pub_date=self.now, is_active=True, headline='Jezdezcon')
-        response = self.client.get(reverse('weblog:index'))
-        self.assertEqual(response.status_code, 200)
-        self.assertQuerysetEqual(response.context['events'], [])
-
-
-class EntryFormTests(TestCase):
-
-    def generate_data(self, rst):
-        return {
-            'headline': 'TestEntry',
-            'slug': 'testentry',
-            'content_format': 'reST',
-            'summary': rst,
-            'body': rst,
-            'pub_date': now(),
-            'author': 'Django developer',
-        }
-
-    def test_validation_of_rst_errors(self):
-        form = EntryForm(self.generate_data('.. tag::'))
-        self.assertFalse(form.is_valid())
-        self.assertIn('summary', form.errors)
-        self.assertEqual(
-            form.errors['summary'][0],
-            'reStructuredText error at line 1: Unknown directive type &quot;tag&quot;.'
-        )
-        self.assertIn('body', form.errors)
-        self.assertEqual(
-            form.errors['body'][0],
-            'reStructuredText error at line 1: Unknown directive type &quot;tag&quot;.'
-        )
-
-    def test_validation_of_rst_warnings(self):
-        form = EntryForm(self.generate_data('Heading\n===='))
-        form.is_valid()
-        self.assertFalse(form.is_valid())
-        self.assertIn('summary', form.errors)
-        self.assertEqual(
-            form.errors['summary'][0],
-            'reStructuredText warning at line 2: Title underline too short.'
-        )
-        self.assertIn('body', form.errors)
-        self.assertEqual(
-            form.errors['body'][0],
-            'reStructuredText warning at line 2: Title underline too short.'
-        )
