+++ blog/tests/test_views.py
@@ -1,0 +1,18 @@
+from django.test import TestCase
+from django_hosts import reverse
+
+from . import DateTimeMixin
+from ..models import Entry, Event
+
+
+class ViewsTestCase(DateTimeMixin, TestCase):
+    def test_no_past_upcoming_events(self):
+        """
+        Make sure there are no past event in the "upcoming events" sidebar (#399)
+        """
+        # We need a published entry on the index page so that it doesn't return a 404
+        Entry.objects.create(pub_date=self.yesterday, is_active=True)
+        Event.objects.create(date=self.yesterday, pub_date=self.now, is_active=True, headline='Jezdezcon')
+        response = self.client.get(reverse('weblog:index'))
+        self.assertEqual(response.status_code, 200)
+        self.assertQuerysetEqual(response.context['events'], [])
